requirements.txt
# Core dependencies
fastapi>=0.110.0
uvicorn>=0.27.0
python-dotenv>=1.0.1
aiohttp>=3.9.3
pydantic>=2.7.4

# Tavily API client
tavily-python>=0.3.0

# AI Model integration
openai>=1.12.0
google-generativeai>=0.4.0

# MCP Server dependencies
websockets>=12.0
jsonrpc>=3.0.0
sse-starlette>=2.0.0

# Additional utilities
httpx>=0.27.0
python-multipart>=0.0.9
anyio>=4.3.0
protobuf>=4.25.3
cryptography>=42.0.0